Refactor odo tree objects hierarchy
