+++ src/openshift/catalog.ts
@@ -3,7 +3,8 @@
  *  Licensed under the MIT License. See LICENSE file in the project root for license information.
  *-----------------------------------------------------------------------------------------------*/
 
-import { Odo, OdoImpl, Command } from '../odo';
+import { Odo, OdoImpl } from '../odo';
+import { Command } from "../odo/command";
 import { Platform } from '../util/platform';
 import { vsCommand } from '../vscommand';
