+++ src/openshift/openshiftItem.ts
@@ -3,9 +3,9 @@
  *  Licensed under the MIT License. See LICENSE file in the project root for license information.
  *-----------------------------------------------------------------------------------------------*/
 
-import { window, QuickPickItem, TreeItemCollapsibleState } from 'vscode';
+import { window, QuickPickItem } from 'vscode';
 import * as validator from 'validator';
-import { Odo, OdoImpl, OpenShiftObject, OpenShiftObjectImpl, ContextType } from '../odo';
+import { Odo, OdoImpl, OpenShiftObject, ContextType, OpenShiftApplication } from '../odo';
 import { OpenShiftExplorer } from '../explorer';
 
 const errorMessage = {
@@ -135,7 +135,7 @@ export default class OpenShiftItem {
             if (context && isCommand(context)) {
                 const newAppName = await context.command();
                 if (newAppName) {
-                    context = new OpenShiftObjectImpl(project, newAppName, ContextType.APPLICATION, false, OdoImpl.Instance, TreeItemCollapsibleState.Collapsed);
+                    context = new OpenShiftApplication(project, newAppName);
                 } else {
                     context = null;
                 }
