+++ test/unit/openshift/service.test.ts
@@ -7,7 +7,8 @@ import * as vscode from 'vscode';
 import * as chai from 'chai';
 import * as sinonChai from 'sinon-chai';
 import * as sinon from 'sinon';
-import { OdoImpl, Command, ContextType } from '../../../src/odo';
+import { OdoImpl, ContextType } from '../../../src/odo';
+import { Command } from "../../../src/odo/command";
 import { TestItem } from './testOSItem';
 import { Progress } from '../../../src/util/progress';
 import { Service } from '../../../src/openshift/service';
