+++ test/unit/openshift/storage.test.ts
@@ -8,7 +8,8 @@ import * as chai from 'chai';
 import * as sinonChai from 'sinon-chai';
 import * as sinon from 'sinon';
 import { TestItem } from './testOSItem';
-import { OdoImpl, Command, ContextType } from '../../../src/odo';
+import { OdoImpl, ContextType } from '../../../src/odo';
+import { Command } from "../../../src/odo/command";
 import { Storage } from '../../../src/openshift/storage';
 import OpenShiftItem from '../../../src/openshift/openshiftItem';
