+++ test/unit/openshift/url.test.ts
@@ -8,7 +8,8 @@ import * as chai from 'chai';
 import * as sinonChai from 'sinon-chai';
 import * as sinon from 'sinon';
 import { TestItem } from './testOSItem';
-import { OdoImpl, Command, ContextType } from '../../../src/odo';
+import { OdoImpl, ContextType } from '../../../src/odo';
+import { Command } from "../../../src/odo/command";
 import { Url } from '../../../src/openshift/url';
 import OpenShiftItem from '../../../src/openshift/openshiftItem';
